测试

+++ include/_joystick.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-extern uint16_t XValue, YValue;
+extern uint32_t XValue, YValue;
 extern _Bool SWState;
 
 void JoyInit(void);
